Move get_bgp_speaker_runningconfig to common (#15528)
What is the motivation for this PR?
In script test_bgp_dual_asn.py, there is an import from the folder tests/generic_config_updater. To minimize cross-module dependencies, we have refactored this function to a common location.

How did you do it?
How did you verify/test it?
