+++ tests/generic_config_updater/test_bgp_speaker.py
@@ -7,6 +7,7 @@ from tests.common.helpers.assertions import pytest_assert
 from tests.common.gu_utils import apply_patch, expect_op_success
 from tests.common.gu_utils import generate_tmpfile, delete_tmpfile
 from tests.common.gu_utils import create_checkpoint, delete_checkpoint, rollback_or_reload
+from tests.common.gu_utils import get_bgp_speaker_runningconfig
 
 pytestmark = [
     pytest.mark.topology('t0'),     # BGP Speaker is limited to t0 only
@@ -54,27 +55,6 @@ def lo_intf_ips(rand_selected_dut, tbinfo):
         if ip and ipv6:
             return ip, ipv6
     pytest_assert(True, "Required ipv4 and ipv6 to start the test")
-
-
-def get_bgp_speaker_runningconfig(duthost):
-    """ Get bgp speaker config that contains src_address and ip_range
-
-    Sample output in t0:
-    ['\n neighbor BGPSLBPassive update-source 10.1.0.32',
-     '\n neighbor BGPVac update-source 10.1.0.32',
-     '\n bgp listen range 10.255.0.0/25 peer-group BGPSLBPassive',
-     '\n bgp listen range 192.168.0.0/21 peer-group BGPVac']
-    """
-    cmds = "show runningconfiguration bgp"
-    output = duthost.shell(cmds)
-    pytest_assert(not output['rc'], "'{}' failed with rc={}".format(cmds, output['rc']))
-
-    # Sample:
-    # neighbor BGPSLBPassive update-source 10.1.0.32
-    # bgp listen range 192.168.0.0/21 peer-group BGPVac
-    bgp_speaker_pattern = r"\s+neighbor.*update-source.*|\s+bgp listen range.*"
-    bgp_speaker_config = re.findall(bgp_speaker_pattern, output['stdout'])
-    return bgp_speaker_config
 
 
 @pytest.fixture(autouse=True)
